Mtp3UserPart interface update
git-svn-id: https://mobicents.googlecode.com/svn/branches/protocols/ss7/1.x@20846 bf0df8d0-2c1f-0410-b170-bd30377b63dc

+++ isup/isup-api/src/main/java/org/mobicents/protocols/ss7/isup/message/ISUPMessage.java
@@ -39,9 +39,9 @@ public interface ISUPMessage {
 	 * Sets sls to be used when this message is send. If message is received, it contians sls of link on which it has been received.
 	 * @param sls
 	 */
-	public void setSls(byte sls);
+	public void setSls(int sls);
 	
-	public byte getSls();
+	public int getSls();
 	/**
 	 * Get mandatory field, CIC.
 	 * @return
